Removed duplication from TestPut
